feat(2202): add Fable 5.1, Mythos 5.1, GPT-6 Astra and GPT-5.6 Cyber
The bundled catalogue stopped at Fable 5 and GPT-5.6 Sol, so the four models
the vendors shipped since then could not be named at all: `--model fable-5-1`
was passed through unmapped and `--model gpt-6-astra` reached Codex as an
unknown slug.

Added, with the vendor-quoted specifications in
docs/case-studies/issue-2202/data/research/online-research.md:

  - `claude-fable-5-1` and `claude-mythos-5-1`, both with 1M context. Anthropic
    moved Fable 5 to its legacy list when 5.1 shipped, so the bare `fable`
    alias follows the vendor to 5.1 — the same precedent as `opus` → Opus 5 in
    issue #2096 — while `fable-5` stays pinned to `claude-fable-5` for anyone
    who asked for that exact generation.
  - `gpt-6-astra`, plus `gpt-5.6-cyber` and the two Daybreak aliases the
    installed Codex CLI already advertises. `getCodexModelVariants()` derives
    the `openai/…` and `openai.…` forms, so no variant is spelled twice.

`defaultModels` is deliberately unchanged. GPT-6 Astra is 2.5x the price of
`gpt-5.6-sol` and preview-gated, and Fable 5.1 is $10/$50 against Opus 5's
$5/$25: making either the default would change every user's bill without being
asked. Both get `defaultFallbackModels` entries instead, so a capacity or
entitlement failure degrades within the family rather than failing the run.

GPT-6 Astra's 1,050,000-token window is *not* recorded in
MODELS_SUPPORTING_1M_CONTEXT. That list exists only to drive the `[1m]` suffix
and `supports1mContext()` returns false for every tool but claude, so an entry
there would be inert and would imply a suffix Codex does not accept.

The additions pushed src/models/index.mjs past the 1350-line warning threshold
of scripts/check-file-line-limits.sh, which tests/extracted-modules-2198.test.mjs
enforces. The bundled catalogue moves to a new leaf module src/models/catalog.mjs
and is re-exported, so the public import surface is unchanged.

Reproducing test: tests/test-issue-2202-new-model-support.mjs, which fails at
e062446e and passes here.

Refs #2202
