+++ tests/test-codex-support.mjs
@@ -94,7 +94,9 @@ test('Codex validates hidden codex-auto-review model id from CLI catalog', () =>
 
 test('Codex primary model names prioritize gpt-5.6-sol and current visible catalog entries', () => {
   // Issue #2027: gpt-5.6-sol leads the primary catalog, with gpt-5.5 kept as the stable fallback.
-  assert.deepEqual(primaryModelNames.codex, ['gpt-5.6-sol', 'gpt-5.5', 'gpt-5.6-terra', 'gpt-5.6-luna', 'gpt-5.4', 'gpt-5.4-mini', 'gpt-5.3-codex-spark', 'formal-ai']);
+  // Issue #2202: GPT-6 Astra is listed second — it is the newest flagship, but it is
+  // preview-gated and 2.5x the price, so it does not take the lead from the default.
+  assert.deepEqual(primaryModelNames.codex, ['gpt-5.6-sol', 'gpt-6-astra', 'gpt-5.5', 'gpt-5.6-terra', 'gpt-5.6-luna', 'gpt-5.4', 'gpt-5.4-mini', 'gpt-5.3-codex-spark', 'formal-ai']);
   assert.equal(primaryModelNames.codex.includes('codex-auto-review'), false);
 });
